+++ components/com_kinoarhiv/views/genres/tmpl/default.php
@@ -12,9 +12,11 @@ defined('_JEXEC') or die;
 ?>
 <div class="uk-article ka-content">
 	<div class="genre-list">
-	<?php for ($i=0, $n=count($this->items); $i<$n; $i++):
+	<?php $items = count($this->items);
+
+	for ($i = 0; $i < $items; $i++):
 		$item = $this->items[$i]; ?>
-		<div class="genre-item"><a href="<?php echo JRoute::_('index.php?option=com_kinoarhiv&view=movies&filters[movies][genre][]='.$item->id.'&Itemid='.$this->itemid); ?>"><?php echo ucfirst($item->name); ?></a> (<?php echo $item->stats; ?>)</div>
+		<div class="genre-item"><a href="<?php echo JRoute::_('index.php?option=com_kinoarhiv&view=movies&filters[movies][genre][]=' . $item->id . '&Itemid=' . $this->itemid); ?>"><?php echo ucfirst($item->name); ?></a> (<?php echo $item->stats; ?>)</div>
 	<?php endfor; ?>
 	</div>
 	<div class="clear"></div>