+++ CHANGELOG.md
@@ -41,7 +41,7 @@
 * Improve display title on some views for some materials.
 * Fixed errors with undefined variables and array keys in autocomplete fields when the ID of the item isn't defined.
 * Update VideoJS to 4.12.15
-* Update Mediaelement.js to 2.19.0
+* Update Mediaelement.js to 2.20.1
 * Fixed bug with undefined variale which may have unexpected results when the user try to delete person from 'cast & crew' table.
 * Added "Links to buy".
 * Added posibility to edit a music genres.